docker/keras-color-svc/requirements.txt
flask
scipy==1.1.0
matplotlib==2.2.2
PILLOW==5.1.0
tensorflow==1.8.0
